# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "又东风小程序后台 Api 文档",
        "description": "又东风小程序后台 Api 文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/admin-api/admin",
            "description": "普通接口地址"
        },
        {
            "url": "/admin-api",
            "description": "通用接口地址"
        }
    ],
    "paths": {
...
        "/keyword": {
            "get": {
                "tags": [
                    "热搜"
                ],
                "summary": "热搜列表",
                "description": "返回热搜列表",
                "operationId": "App\\Http\\Controllers\\Admin\\SearchHotKeywordsController::index",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/current_page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "关键词",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "添加类型[1:系统添加 2:后台添加]",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_show",
                        "in": "query",
                        "description": "是否展示[0:不展示 1:展示]",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_hot",
                        "in": "query",
                        "description": "是否热门[0:非热门 1:热门]",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_del",
                        "in": "query",
                        "description": "是否删除[0:正常 1:删除]",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
        "/keyword/{keyword}": {
            "get": {
                "tags": [
                    "热搜"
                ],
                "summary": "返回热搜词",
                "description": "根据ID查询热搜词",
                "operationId": "App\\Http\\Controllers\\Admin\\SearchHotKeywordsController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/keyword_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "delete": {
                "tags": [
                    "热搜"
                ],
                "summary": "返回删除状态",
                "description": "根据ID删除热搜",
                "operationId": "App\\Http\\Controllers\\Admin\\SearchHotKeywordsController::destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/keyword_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
        "/ship": {
            "get": {
                "tags": [
                    "配送方式"
                ],
                "summary": "配送方式列表",
                "description": "返回配送方式列表",
                "operationId": "App\\Http\\Controllers\\Admin\\ShipController::index",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/current_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/responses/ship"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "post": {
                "tags": [
                    "配送方式"
                ],
                "summary": "保存配送方式",
                "description": "返回保存的配送方式",
                "operationId": "App\\Http\\Controllers\\Admin\\ShipController::store",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ship_in_body"
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
        "/ship/{ship}": {
            "get": {
                "tags": [
                    "配送方式"
                ],
                "summary": "通过ID查询配送方式",
                "description": "返回查询到的配送方式",
                "operationId": "App\\Http\\Controllers\\Admin\\ShipController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ship_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "delete": {
                "tags": [
                    "配送方式"
                ],
                "summary": "通过ID删除配送方式",
                "description": "返回是否删除成功",
                "operationId": "App\\Http\\Controllers\\Admin\\ShipController::destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ship_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "patch": {
                "tags": [
                    "配送方式"
                ],
                "summary": "通过ID更新配送方式",
                "description": "返回更新后的配送方式",
                "operationId": "App\\Http\\Controllers\\Admin\\ShipController::update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ship_in_path_required"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/ship_in_body"
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
...
        "/spec/{spec}": {
            "get": {
                "tags": [
                    "属性"
                ],
                "summary": "返回ID所属属性",
                "description": "通过ID查询属性",
                "operationId": "App\\Http\\Controllers\\Admin\\SpecController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spec_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/responses/spec"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "delete": {
                "tags": [
                    "属性"
                ],
                "summary": "返回状态",
                "description": "删除属性",
                "operationId": "App\\Http\\Controllers\\Admin\\SpecController::destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spec_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "patch": {
                "tags": [
                    "属性"
                ],
                "summary": "返回更新的属性详情",
                "description": "更新属性",
                "operationId": "App\\Http\\Controllers\\Admin\\SpecController::update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spec_in_path_required"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/spec_in_body"
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/responses/spec"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ship": {
                "title": "配送方式模板",
                "description": "配送方式模板",
                "properties": {
                    "name": {
                        "description": "配送方式名称",
                        "format": "string",
                        "example": "配送方式1"
                    },
                    "type": {
                        "description": "计算方式[1:按重量 2:按件数]",
                        "format": "integer",
                        "example": 1
                    },
                    "has_cod": {
                        "description": "是否货到付款[1:不是 2:是]",
                        "format": "integer",
                        "example": 1
                    },
                    "firstunit": {
                        "description": "首重(单位:克)",
                        "format": "integer",
                        "example": 500
                    },
                    "continueunit": {
                        "description": "续重(单位:克)",
                        "format": "integer",
                        "example": 500
                    },
                    "def_area_fee": {
                        "description": "按地区设置配送费用是否启用默认配送费用[1:启用 2:不启用]",
                        "format": "integer",
                        "example": 1
                    },
                    "area_type": {
                        "description": "地区类型[1:全部地区 2:指定地区]",
                        "format": "integer",
                        "example": 1
                    },
                    "firstunit_price": {
                        "description": "首重费用",
                        "format": "float",
                        "example": 10
                    },
                    "continueunit_price": {
                        "description": "续重费用",
                        "format": "float",
                        "example": 10
                    },
                    "logi_name": {
                        "description": "物流公司名称",
                        "format": "string",
                        "example": "顺丰速运"
                    },
                    "logi_code": {
                        "description": "物流公司编码",
                        "format": "string",
                        "example": "顺丰速运"
                    },
                    "is_def": {
                        "description": "是否默认[1:默认 2:不默认]",
                        "format": "integer",
                        "example": 2
                    },
                    "sort": {
                        "description": "排序",
                        "format": "integer",
                        "example": 100
                    },
                    "status": {
                        "description": "状态[1:正常 2:禁用]",
                        "format": "integer",
                        "example": 1
                    },
                    "free_postage": {
                        "description": "是否包邮[1:包邮 2:不包邮]",
                        "format": "integer",
                        "example": 2
                    },
                    "goodsmoney": {
                        "description": "满多少免运费",
                        "format": "float",
                        "example": 200
                    },
                    "area_fee": {
                        "description": "地区配送费用",
                        "format": "string",
                        "example": {
                            "area_id": [
                                10005,
                                2005,
                                25576
                            ],
                            "firstunit": 500,
                            "continueunit": 500,
                            "firstunit_price": 12,
                            "continueunit_price": 8
                        }
                    }
                },
                "type": "object"
            },
            "Spec": {
                "title": "属性模型",
                "description": "属性模型",
                "properties": {
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "内存"
                    },
                    "sort": {
                        "type": "integer",
                        "format": "integer",
                        "example": 100
                    },
                    "details": {
                        "type": "string",
                        "format": "string",
                        "nullable": true
                    },
                    "values": {
                        "description": "@OA\\Property(format=\"string\",example={34,35})",
                        "format": "string",
                        "example": [
                            34,
                            35
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "spec": {
                "description": "属性信息",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Spec"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "current_page": {
                "name": "current_page",
                "in": "path",
                "description": "当前页",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64",
                    "default": 1
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "path",
                "description": "每页显示数量",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64",
                    "default": 10
                }
            },
            "keyword_in_path_required": {
                "name": "keyword",
                "in": "path",
                "description": "热搜ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "ship_in_path_required": {
                "name": "ship",
                "in": "path",
                "description": "配送方式ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "spec_in_path_required": {
                "name": "spec",
                "in": "path",
                "description": "属性ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "requestBodies": {
            "ship_in_body": {
                "description": "配送方式表单数据",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Ship"
                        }
                    }
                }
            },
            "spec_in_body": {
                "description": "属性表单数据",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Spec"
                        }
                    }
                }
            },
...
        }
    }
}